requirements.txt
streamlit>=1.28.0
discord.py>=2.3.0
langchain>=0.0.350
langchain-community>=0.0.350
langchain-openai>=0.0.350
chromadb>=0.4.0
PyPDF2>=3.0.0
pycryptodome>=3.23.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
openai>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0